deployer.sh changed for static files

+++ deployer.sh
@@ -9,16 +9,20 @@ case $1 in
    "deploy")
                 echo "deployer.deploy"
                 cd $2
-                needCI=$(git diff --stat master@{1} master package-lock.json || echo "err")
-                if [[ $needCI = "" ]]
-                  then echo "only build"
-                  npm run build
+                if [ -f ./package.json ]; then
+                  needCI=$(git diff --stat master@{1} master package-lock.json || echo "err")
+                  if [ $needCI = "" ]; then
+                    echo "only build"
+                    npm run build
+                  else
+                    echo "ci and build"
+                    npm ci
+                    npm run build
+                  fi
+                  cp -r ./build /usr/share/nginx/html
                 else
-                  echo "ci and build"
-                  npm ci
-                  npm run build
+                  cp ./* /usr/share/nginx/html
                 fi
-                cp -r ./build /usr/share/nginx/html
                 service nginx stop || true
                 service nginx stop
                 service nginx start
